[REF] l10n_br_fiscal: Merge document classes
